requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
fastapi==0.55.1
fastapi-contrib==0.2.6
h11==0.9.0
httptools==0.1.1
idna==2.9
multidict==4.7.6
numpy==1.19.0
pandas==1.0.5
passlib==1.7.2
psycopg2-binary==2.8.5
pydantic==1.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
starlette==0.13.2
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
